refactor wallet to its own package
