+++ src/cryptoadvance/specter/wallet/__init__.py
@@ -1,0 +1,1 @@
+from .wallet import Wallet, purposes
